Move build targets to targets/
